requirements.txt
aiohttp==3.5.4
async-timeout==3.0.0
attrs==18.1.0
boto3==1.12.37
botocore==1.15.37
cairocffi==1.1.0
CairoSVG==2.4.2
chardet==3.0.4
cssselect2==0.3.0
defusedxml==0.6.0
discord.py==1.3.4
docutils==0.15.2
fuzzywuzzy==0.16.0
idna==2.7
jmespath==0.9.5
lxml==4.5.0
multidict==4.5.2
numpy==1.14.5
Pillow==7.1.1
pycparser==2.18
PyNaCl==1.3.0
python-dateutil==2.8.1
python-Levenshtein==0.12.0
s3transfer==0.3.3
six==1.11.0
svgutils==0.3.1
tinycss2==1.0.2
urllib3==1.25.8
uuid==1.30
webencodings==0.5.1
websockets==6.0
yarl==1.3.0
psycopg2==2.8.6
